% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotation.R
\name{cotation}
\alias{cotation}
\title{Show prices from Tunis Stock Exchange}
\usage{
cotation(x)
}
\arguments{
\item{"x"}{the ticker symbol of firm}
}
\value{
Historical Prices Graph
}
\description{
Show prices from Tunis Stock Exchange
}
\examples{
cotation("STB")
}
