% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Prices of Tunisian firms}
\format{
A data frame with 407199 rows and 3 variables:
\describe{
  \item{seance}{date of cotation, daily}
  \item{val}{ticker, in carats}
  \item{p}{price, in tunisian dinars}
}
}
\source{
\url{https://tinyurl.com/fouedhamouda}
}
\usage{
sample_data
}
\description{
A dataset containing the prices of Tunsisian firms
}
\keyword{datasets}
