% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotation.R
\name{cotation}
\alias{cotation}
\title{Tunisian firm cotation}
\usage{
cotation(x)
}
\arguments{
\item{x}{the ticker symbol of firm}
}
\value{
Historical Prices Graph
}
\description{
Get prices from Tunis Stock Exchange
}
